test: shift income junit tests from ParserTest to IncomeTest
